refactor: DraggableResizableコンポーネントをTimelinePanelから分離し汎用化
- ドラッグ＆リサイズ機能を再利用可能な独立したコンポーネントとして実装
- TypeScriptの型定義を追加し型安全性を向上
- コンポーネントの責務を明確に分離しSRP原則に準拠
